FIX: Compile warnings and 64bit static code analyzer warnings has been fixed. FIX: Many minor bugs fixed.
svn path=/trunk/; revision=832

+++ Source/ZEEngine/ZECore/ZETimerManager.cpp
@@ -38,7 +38,7 @@
 
 ZETimerManager::ZETimerManager()
 {
-	CurrentTime = ZERealTimeClock::GetInstance()->GetCurrentTime();
+	CurrentTime = (float)ZERealTimeClock::GetInstance()->GetCurrentTime();
 }
 
 ZETimerManager::~ZETimerManager()
